Use same request wrapper for both odata providers
