+++ .idea/runConfigurations/Run_Performance_Tests.xml
@@ -8,7 +8,7 @@
     </extension>
     <module name="nuget-tests" />
     <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-    <option name="ALTERNATIVE_JRE_PATH" value="" />
+    <option name="ALTERNATIVE_JRE_PATH" />
     <option name="SUITE_NAME" value="$PROJECT_DIR$/nuget-tests/src/testng-nuget-performance.xml" />
     <option name="PACKAGE_NAME" value="" />
     <option name="MAIN_CLASS_NAME" value="" />
