+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/controllers/NuGetFeedHandler.java
@@ -27,9 +27,7 @@ import java.io.IOException;
  *         Date: 17.01.12 17:46
  */
 public interface NuGetFeedHandler {
-  boolean isAvailable();
 
-  void handleRequest(@NotNull final String baseMappingPath,
-                     @NotNull final HttpServletRequest request,
+  void handleRequest(@NotNull final HttpServletRequest request,
                      @NotNull final HttpServletResponse response) throws Exception;
 }
